SR Class/Readme (#2)
Overview
I was unable to figure out why the title attribute did not contribute to the reading of the page on an SR, but quickly found a substitute.

Updated the README to better suit the project.

HTML/CSS
Added the SR class.

Javascript
n/a

Repository
Finished the README, tested with NVDA SR program.